change getStayPoints to not update before running, remove algorithm parameters from persisted staypoints
